Crate polycool

This is a crate for numerical polynomial root-finding.

Currently, we implement a single solver: Yuksel’s iterative algorithm for finding roots in a bounded interval. We aspire to have more, with thorough tests and benchmarks.

Modules

cubic 🔒

poly 🔒

poly_dyn 🔒

Polynomials of dynamic (run-time) degree.

quadratic 🔒

yuksel 🔒

An implementation of Yuksel’s robust version of Newton’s algorithm.

Structs

Poly

A polynomial whose degree is known at compile-time.

PolyDyn

A polynomial of dynamic degree.

Functions

different_signs 🔒

Type Aliases

Cubic

A polynomial of degree 3.

Quadratic

A polynomial of degree 2.

Quartic

A polynomial of degree 4.

Quintic

A polynomial of degree 5.